NLP Profiling + MetricSpec related fixes (#11)
* NLP Profile

* Fix text profiling

* Edit MetricSpec endpoint

* Fix time for metrics

* Filter metrics with MetricSpec id

* Fix trace serialization + return origin traces

+++ src/main/scala/io/hydrosphere/sonar/terms/PreprocessedProfile.scala
@@ -28,7 +28,7 @@ case class TextPreprocessedProfile(
   lengthSum: Long,
   tokenLengthSum: Long,
   treeDepthSum: Long,
-  uniqueLemmasSum: Long,
+  uniqueLemmasSum: Double,
   languagesSum: Map[String, Long],
   posTagsSum: Map[String, Long],
   tokenHyperLogLog: HyperLogLog
